CPI datasets and integration tests that load them. Std COICOP and SpecAgg versions for Dec 2016.
