+++ it/main/LoadDataUsingNewDimensionsTest.java
@@ -12,15 +12,9 @@ import uk.co.onsdigital.discovery.model.DimensionalDataSet;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
 import javax.persistence.EntityTransaction;
-import java.io.BufferedWriter;
-import java.io.File;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
 import static main.PostgresTest.*;
 import static org.testng.Assert.assertEquals;
@@ -145,74 +139,6 @@ public class LoadDataUsingNewDimensionsTest extends TestNGSuite {
                 tx.rollback();
             }
         });
-    }
-
-
-    @Test(enabled = false)
-    public void convertDimensionInFile() throws Exception {
-
-        int[] tripletStartindices = new int[]{7, 10};
-        String inputFileName = "Open-Data-v3.csv";
-        String outputFileName = "plop.csv";
-
-        EntityTransaction tx = em.getTransaction();
-        tx.begin();
-        try {
-            postgresTest.loadStandingData(em, Arrays.asList(NACE));
-            postgresTest.loadStandingData(em, Arrays.asList(PRODCOM_ELEMENTS));
-
-
-            for (int i = 0; i < tripletStartindices.length; i++) {
-                int tripletStartindex = tripletStartindices[i];
-
-                File inputFile;
-                File outputFile;
-                BufferedWriter writer;
-
-                if (i == 0) {
-                    inputFile = new File(new PostgresTest().getClass().getResource(inputFileName).getPath());
-                    writer = Files.newBufferedWriter(Paths.get(new File(outputFileName + "_" + i).getAbsolutePath()));
-                } else {
-                    inputFile = new File(new PostgresTest().getClass().getResource(outputFileName + "_" + (i - 1)).getPath());
-                    writer = Files.newBufferedWriter(Paths.get(new File(outputFileName + "_" + i).getAbsolutePath()));
-                }
-
-                ArrayList<String> lines = Files.lines(Paths.get(inputFile.getAbsolutePath())).collect(Collectors.toCollection(ArrayList::new));
-
-                int counter = 0;
-
-                for (String line : lines) {
-                    if (line.isEmpty() || counter == 0) {
-                        writer.write(line + "\n");
-                    } else {
-
-                        String[] lineParts = line.split(",(?=([^\"]*\"[^\"]*\")*[^\"]*$)");
-                        String hierarchyId = lineParts[tripletStartindex];
-                        String dimName = lineParts[tripletStartindex + 1];
-                        String dimValue = lineParts[tripletStartindex + 2].replace("\"", "");
-
-                        logger.debug("Looking up code for hierarchy: " + hierarchyId + " and dimValue: " + dimValue);
-
-                        String convertedDimValue = em.createQuery("SELECT he.code FROM HierarchyEntry he WHERE he.hierarchy.id = :hierarchyId AND he.name = :dimValue", String.class)
-                                .setParameter("hierarchyId", hierarchyId)
-                                .setParameter("dimValue", dimValue)
-                                .getSingleResult();
-
-                        lineParts[tripletStartindex + 2] = convertedDimValue;
-                        String newLine = Arrays.stream(lineParts).collect(Collectors.joining(","));
-                        logger.debug("Outputting ammended line: " + newLine);
-                        writer.write(newLine + "\n");
-                        writer.flush();
-                    }
-                    counter++;
-                }
-            }
-        } catch (Exception e) {
-            e.printStackTrace();
-            fail();
-        } finally {
-            tx.rollback();
-        }
     }
 
 }
