+++ it/resources/main/CPI_2016_12_COICOP_v3.csv
@@ -1,0 +1,1 @@
+Observation,Data_Marking,Observation_Type_Value,Dimension_Hierarchy_1,Dimension_Name_1,Dimension_Value_1,Dimension_Hierarchy_2,Dimension_Name_2,Dimension_Value_2101.9,,,time,Month,2016.12,CL_0000641,COICOP,CI_000421698.1,,,time,Month,2016.12,CL_0000641,COICOP,CI_000421796.3,,,time,Month,2016.12,CL_0000641,COICOP,CI_000421894,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004219106.8,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004220103.1,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004221100.1,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004222102.3,,,time,Month,2016.12,CL_0000641,COICOP,CI_000422399.5,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004224101.2,,,time,Month,2016.12,CL_0000641,COICOP,CI_000422597.5,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004226104,,,time,Month,2016.12,CL_0000641,COICOP,CI_000422799.9,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004228103.7,,,time,Month,2016.12,CL_0000641,COICOP,CI_000422997.1,,,time,Month,2016.12,CL_0000641,COICOP,CI_000423098.7,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004231100.8,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004232101.6,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004233102.9,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004354102.9,,,time,Month,2016.12,CL_0000641,COICOP,CI_000435598.3,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004234103.2,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004235110.9,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004236101.9,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004237103.3,,,time,Month,2016.12,CL_0000641,COICOP,CI_000423895.3,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004239102.2,,,time,Month,2016.12,CL_0000641,COICOP,CI_000435699.4,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004240106.2,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004241106.2,,,time,Month,2016.12,CL_0000641,COICOP,CI_000424299.5,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004243107.9,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004244103.3,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004245104.2,,,time,Month,2016.12,CL_0000641,COICOP,CI_000424699.1,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004247100.2,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004248105.1,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004357110.3,,,time,Month,2016.12,CL_0000641,COICOP,CI_000424999.5,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004250100.3,,,time,Month,2016.12,CL_0000641,COICOP,CI_000425199.8,,,time,Month,2016.12,CL_0000641,COICOP,CI_000425295.9,,,time,Month,2016.12,CL_0000641,COICOP,CI_000425399.9,,,time,Month,2016.12,CL_0000641,COICOP,CI_000425496.5,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004255102.1,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004256105.6,,,time,Month,2016.12,CL_0000641,COICOP,CI_000425795.2,,,time,Month,2016.12,CL_0000641,COICOP,CI_000425896.6,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004259101.3,,,time,Month,2016.12,CL_0000641,COICOP,CI_000426098.8,,,time,Month,2016.12,CL_0000641,COICOP,CI_000426195.5,,,time,Month,2016.12,CL_0000641,COICOP,CI_000426293.4,,,time,Month,2016.12,CL_0000641,COICOP,CI_000426393.3,,,time,Month,2016.12,CL_0000641,COICOP,CI_000426495.9,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004265106.8,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004266103.2,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004267102.8,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004268103.3,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004269100.1,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004270102.3,,,time,Month,2016.12,CL_0000641,COICOP,CI_000427197.9,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004272101,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004273100.5,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004274102,,,time,Month,2016.12,CL_0000641,COICOP,CI_000427599.7,,,time,Month,2016.12,CL_0000641,COICOP,CI_000427693.3,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004277110.8,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004278101.7,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004279103.3,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004280106.6,,,time,Month,2016.12,CL_0000641,COICOP,CI_000428199.9,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004282103.9,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004283102.4,,,time,Month,2016.12,CL_0000641,COICOP,CI_000428497.1,,,time,Month,2016.12,CL_0000641,COICOP,CI_000428598.7,,,time,Month,2016.12,CL_0000641,COICOP,CI_000428694.7,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004287104.7,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004288102.3,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004289100.5,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004290101.8,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004358104.3,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004359102.9,,,time,Month,2016.12,CL_0000641,COICOP,CI_000436098.3,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004291101.3,,,time,Month,2016.12,CL_0000641,COICOP,CI_000429293.4,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004293101.9,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004294103.1,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004295102.5,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004296103.9,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004297100.6,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004298106.7,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004299119.4,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004300109.9,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004301101.9,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004302103.3,,,time,Month,2016.12,CL_0000641,COICOP,CI_000430395.9,,,time,Month,2016.12,CL_0000641,COICOP,CI_000430492.1,,,time,Month,2016.12,CL_0000641,COICOP,CI_000430589.3,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004306106.4,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004307101.1,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004308102.2,,,time,Month,2016.12,CL_0000641,COICOP,CI_000436198,,,time,Month,2016.12,CL_0000641,COICOP,CI_000430999.8,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004310100.7,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004311101.6,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004312108.1,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004313105.5,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004314108.5,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004315105.6,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004316104.6,,,time,Month,2016.12,CL_0000641,COICOP,CI_000431799.5,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004318107.9,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004319103.5,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004320102.1,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004321104.2,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004322102.8,,,time,Month,2016.12,CL_0000641,COICOP,CI_000432397.9,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004324102.6,,,time,Month,2016.12,CL_0000641,COICOP,CI_000432597.6,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004326105.1,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004362101.3,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004327106.4,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004363116.8,,,time,Month,2016.12,CL_0000641,COICOP,CI_000432899.5,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004329100.3,,,time,Month,2016.12,CL_0000641,COICOP,CI_000433097.9,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004331101.3,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004332102.7,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004333100.7,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004334101.7,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004335102.2,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004336103.7,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004337103.2,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004338101.2,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004339107.9,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004340103.5,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004341101.5,,,time,Month,2016.12,CL_0000641,COICOP,CI_0004342
