+++ it/resources/main/CPI_2016_12_SpecialAggregate_v3.csv
@@ -1,0 +1,1 @@
+Observation,Data_Marking,Observation_Type_Value,Dimension_Hierarchy_1,Dimension_Name_1,Dimension_Value_1,Dimension_Hierarchy_2,Dimension_Name_2,Dimension_Value_2100,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_000515698.9,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_000515797.9,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_000515897.9,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_000515999.6,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005160101.3,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005161100.5,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005163100.5,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_000516496.7,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005165103.6,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005166100.5,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005167102.7,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005168101,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005169101.2,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005170100.1,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005171101.6,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_000517298.7,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_000517399.9,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_000517495,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005175101.5,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_000517698.8,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005177104,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005178102.5,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005179101.7,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005180104.4,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005181107.7,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005182103,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005183110.9,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005184103.2,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005185101.3,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005186104,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005187103.3,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005188105.4,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005189104.1,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005190102.3,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005191102.8,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_000521297.5,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_000519299.6,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005193101.3,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005194100.2,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_000519599.4,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_000519699.4,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005197100.2,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005198105.1,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005199102,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005200102.6,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005201102,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005202102.1,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005203101.8,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005204102.3,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005205101.9,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005206102.1,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005207101.7,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005208101.8,,,time,Month,2016.12,CL_0000641,Special Aggregate,CI_0005209
